fix: comments, refactor code
