+++ src/Context/index.js
@@ -1,9 +1,5 @@
 export { ModalProvider, useModal } from "./Modal/ModalContext";
 export { ScrollToTop } from "./ScrollToTop/ScrollToTop";
-export {
-  useLocalStorageGet,
-  useLocalStorageSet,
-} from "../Hooks/useLocalStorage";
 export { AnimationProvider, useAnimation } from "./Animations/AnimationContext";
 export { AxiosCallProvider, useAxiosCalls } from "./AxiosCalls/AxiosContext";
 export { ThemeProvider, useTheme } from "./Theme/ThemeContext";
