--- src/Hooks/useLocalStorage.jsx
@@ -1,10 +1,0 @@
-import { useEffect } from "react";
-
-const useLocalStorageSet = (key, value) =>
-  useEffect(() => {
-    localStorage.setItem(key, JSON.stringify(value));
-  }, [value, key]);
-
-const useLocalStorageGet = (key) => JSON.parse(localStorage.getItem(key));
-
-export { useLocalStorageGet, useLocalStorageSet };
