+++ src/Pages/Landing/Landing.jsx
@@ -1,22 +1,21 @@
 import { VideoCard, BannerCard, CategoryCard, Footer } from "Components";
 import { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { filterActions } from "Store";
+import { filterActions } from "Store/store";
 import "./Landing.css";
 
 export const Landing = () => {
-  // redux
   const {
     videoState: { videos, categories },
   } = useSelector((videoState) => videoState);
   const dispatch = useDispatch();
 
-  const mapVideos = videos.map((video) => {
-    return video.top && <VideoCard key={video._id} videoDetail={video} />;
+  const mapVideos = videos?.map((video) => {
+    return video?.top && <VideoCard key={video?._id} videoDetail={video} />;
   });
 
-  const mapCategories = categories.map((category) => (
-    <CategoryCard key={category._id} category={category} />
+  const mapCategories = categories?.map((category) => (
+    <CategoryCard key={category?._id} category={category} />
   ));
 
   useEffect(() => {
