+++ src/Pages/Watchlater/Watchlater.jsx
@@ -3,7 +3,6 @@ import { useSelector } from "react-redux";
 import "./Watchlater.css";
 
 export const Watchlater = () => {
-  // redux
   const {
     videoState: { watchlater },
   } = useSelector((videoState) => videoState);
@@ -13,10 +12,10 @@ export const Watchlater = () => {
   } = useSelector((authState) => authState);
 
   const mapWatchlater =
-    watchlater.length > 0 ? (
+    watchlater?.length > 0 ? (
       <div className="watchlater-video-section flex-row">
-        {watchlater.map((video) => (
-          <VideoCard key={video._id} videoDetail={video} />
+        {watchlater?.map((video) => (
+          <VideoCard key={video?._id} videoDetail={video} />
         ))}
       </div>
     ) : (
